Refactory of importSQLFile
 - Method exported to its own file
